delete application pop-up

+++ src/main/webapp/WEB-INF/view/student.jsp
@@ -53,18 +53,41 @@
                 </p>
                 <c:if test="${ status  == 'Pending'}">
                     <a class="btn btn-danger"
-                       href="/delete_application?applicationId=${application.id}"
-                       onclick="return confirm('Withdraw this application?')">Withdraw</a>
+                       href="#"
+                       data-toggle="modal" data-target="#modal_${application.id}">Withdraw</a>
                 </c:if>
             </div>
         </div>
         <br>
+        <!-- Modal -->
+        <div class="modal fade" id="modal_${application.id}" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel"
+             aria-hidden="true">
+            <div class="modal-dialog" role="document">
+                <div class="modal-content">
+                    <div class="modal-header">
+                        <h5 class="modal-title" id="exampleModalLabel">Housing</h5>
+                        <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+                            <span aria-hidden="true">&times;</span>
+                        </button>
+                    </div>
+                    <div class="modal-body">
+                        Withdraw this application?
+                    </div>
+                    <div class="modal-footer">
+                        <button type="button" class="btn btn-secondary" data-dismiss="modal">Cancel</button>
+                        <a class="btn btn-danger"
+                           href="/delete_application?applicationId=${application.id}">Confirm</a>
+                    </div>
+                </div>
+            </div>
+        </div>
     </c:forEach>
 
 
     <!-- footer -->
     <jsp:include page="footer.jsp"></jsp:include>
     <!-- footer -->
+
 
 </div>
 <!-- /container -->
